refactor(message): extract MessageSendService for the outbound send family
MessageService mixed two responsibilities: executing outbound sends
end-to-end (engine resolution, media prep, pending/SENT persistence,
send-path plugin hooks, pacing) and querying/mutating message state.
The send family now lives on MessageSendService, which owns the gate,
failure funnel, outgoing persistence and outbound archiving.

MessageService keeps the query and engine-action mutation surface and
exposes each moved verb as a thin same-signature pass-through, so the
controller, agent tools, plugin capabilities and bulk send keep their
existing wiring. sendOutgoingMessage stays reachable for the bulk
path's persist-after-send flow. editMessage keeps its moderation gate
on the query side, sharing the core/hooks sending-gate implementation.

Hook emissions keep their `source: 'MessageService'` context so the
split is not observable to plugins; the send service logs under the
MessageSend channel. Specs split along the same seam, with a facade
delegation test pinning the pass-through contract.
