+++ src/modules/message/send-pacing.service.spec.ts
@@ -15,7 +15,7 @@ import { MessageDirection, type Message } from './entities/message.entity';
 import type { Session } from '../session/entities/session.entity';
 import type { Repository } from 'typeorm';
 import type { ConfigService } from '@nestjs/config';
-import { MessageService } from './message.service';
+import { MessageSendService } from './message-send.service';
 import { BulkMessageService } from './bulk-message.service';
 import { StatusService } from '../status/status.service';
 import { CatalogService } from '../catalog/catalog.service';
@@ -299,17 +299,15 @@ describe('send paths consult the governor', () => {
     assertSendAllowed: jest.fn().mockRejectedValue(new HttpException({ code: SEND_PACING_LIMITED }, 429)),
   });
 
-  it('MessageService refuses before the plugin gate runs and before the engine is asked', async () => {
+  it('MessageSendService refuses before the plugin gate runs and before the engine is asked', async () => {
     const hookManager = { execute: jest.fn() };
     const engine = { sendTextMessage: jest.fn() };
     const pacing = refusing();
-    const service = new MessageService(
+    const service = new MessageSendService(
       { save: jest.fn() } as never,
       { isActive: () => true } as never,
       { get: () => engine, require: () => engine } as never,
-      {} as never,
       hookManager as never,
-      {} as never,
       {} as never,
       pacing as never,
     );
